+++ lang/elaborator/src/typechecker/exprs/typecheck.rs
@@ -2,7 +2,6 @@
 
 use std::rc::Rc;
 
-use askama::Template;
 use codespan::Span;
 use log::trace;
 
@@ -19,383 +18,8 @@ use syntax::ctx::{BindContext, BindElem, LevelCtx};
 
 use super::super::ctx::*;
 use super::super::util::*;
+use super::CheckInfer;
 use crate::result::TypeError;
-
-/// The CheckInfer trait for bidirectional type inference.
-/// Expressions which implement this trait provide both a `check` function
-/// to typecheck the expression against an expected type and a `infer` function
-/// to infer the type of the given expression.
-pub trait CheckInfer: Sized {
-    /// Checks whether the expression has the given expected type. For checking we use
-    /// the following syntax:
-    /// ```text
-    ///            P, Γ ⊢ e ⇐ τ
-    /// ```
-    /// - P: The program context of toplevel declarations.
-    /// - Γ: The context of locally bound variables.
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError>;
-    /// Tries to infer a type for the given expression. For inference we use the
-    /// following syntax:
-    /// ```text
-    ///            P, Γ ⊢ e ⇒ τ
-    /// ```
-    ///  - P: The program context of toplevel declarations.
-    ///  - Γ: The context of locally bound variables.
-    fn infer(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError>;
-}
-
-impl<T: CheckInfer> CheckInfer for Rc<T> {
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        Ok(Rc::new((**self).check(prg, ctx, t)?))
-    }
-    fn infer(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
-        Ok(Rc::new((**self).infer(prg, ctx)?))
-    }
-}
-
-// Expressions
-//
-//
-
-impl CheckInfer for Exp {
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        trace!(
-            "{} |- {} <= {}",
-            ctx.print_to_string(None),
-            self.print_to_string(None),
-            t.print_to_string(None)
-        );
-        match self {
-            Exp::Variable(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::TypCtor(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::Call(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::DotCall(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::Anno(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::TypeUniv(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::Hole(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::LocalMatch(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-            Exp::LocalComatch(e) => Ok(e.check(prg, ctx, t.clone())?.into()),
-        }
-    }
-
-    fn infer(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
-        let res: Result<Exp, TypeError> = match self {
-            Exp::Variable(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::TypCtor(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::Call(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::DotCall(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::Anno(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::TypeUniv(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::Hole(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::LocalMatch(e) => Ok(e.infer(prg, ctx)?.into()),
-            Exp::LocalComatch(e) => Ok(e.infer(prg, ctx)?.into()),
-        };
-        trace!(
-            "{} |- {} => {}",
-            ctx.print_to_string(None),
-            self.print_to_string(None),
-            res.as_ref().map(|e| e.typ()).print_to_string(None)
-        );
-        res
-    }
-}
-
-// Variable
-//
-//
-
-#[derive(Template)]
-#[template(path = "var_rule.txt")]
-struct VarRule<'a> {
-    ctx: &'a str,
-    var: &'a str,
-    typ: &'a str,
-}
-
-impl CheckInfer for Variable {
-    /// The *checking* rule for variables is:
-    /// ```text
-    ///            P, Γ ⊢ x ⇒ τ
-    ///            P, Γ ⊢ τ ≃ σ
-    ///           ───────────────
-    ///            P, Γ ⊢ x ⇐ σ
-    /// ```
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        let inferred_term = self.infer(prg, ctx)?;
-        let inferred_typ = inferred_term.typ().ok_or(TypeError::Impossible {
-            message: "Expected inferred type".to_owned(),
-            span: None,
-        })?;
-        convert(ctx.levels(), &mut ctx.meta_vars, inferred_typ, &t)?;
-        Ok(inferred_term)
-    }
-
-    /// The *inference* rule for variables is:
-    /// ```text
-    ///            Γ(x) = τ
-    ///           ───────────────
-    ///            P, Γ ⊢ x ⇒ τ
-    /// ```
-    fn infer(&self, _prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
-        let Variable { span, idx, name, .. } = self;
-        let typ_nf = ctx.lookup(*idx);
-
-        trace!(
-            "{}",
-            VarRule {
-                ctx: &ctx.vars.print_to_string(None),
-                var: name,
-                typ: &typ_nf.print_to_string(None)
-            }
-            .render()
-            .unwrap()
-        );
-
-        Ok(Variable { span: *span, idx: *idx, name: name.clone(), inferred_type: Some(typ_nf) })
-    }
-}
-
-// TypCtor
-//
-//
-
-impl CheckInfer for TypCtor {
-    /// The *checking* rule for type constructors is:
-    /// ```text
-    ///            P, Γ ⊢ Tσ ⇒ ρ
-    ///            P, Γ ⊢ τ ≃ ρ
-    ///           ──────────────────
-    ///            P, Γ ⊢ Tσ ⇐ τ
-    /// ```
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        let inferred_term = self.infer(prg, ctx)?;
-        let inferred_typ = inferred_term.typ().ok_or(TypeError::Impossible {
-            message: "Expected inferred type".to_owned(),
-            span: None,
-        })?;
-        convert(ctx.levels(), &mut ctx.meta_vars, inferred_typ, &t)?;
-        Ok(inferred_term)
-    }
-
-    /// The *inference* rule for type constructors is:
-    /// ```text
-    ///            (co)data Tψ {...} ∈ P
-    ///            P, Γ ⊢ σ ⇐ ψ
-    ///           ─────────────────────────
-    ///            P, Γ ⊢ Tσ ⇒ Type
-    /// ```
-    fn infer(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
-        let TypCtor { span, name, args } = self;
-        let params = &*prg.typ(name, *span)?.typ();
-        let args_out = check_args(args, prg, name, ctx, params, *span)?;
-
-        Ok(TypCtor { span: *span, name: name.clone(), args: args_out })
-    }
-}
-
-// Call
-//
-//
-
-impl CheckInfer for Call {
-    /// The *checking* rule for calls is:
-    /// ```text
-    ///                 ...
-    ///           ──────────────────
-    ///            P, Γ ⊢ Cσ ⇐ τ
-    /// ```
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        let inferred_term = self.infer(prg, ctx)?;
-        let inferred_typ = inferred_term.typ().ok_or(TypeError::Impossible {
-            message: "Expected inferred type".to_owned(),
-            span: None,
-        })?;
-        convert(ctx.levels(), &mut ctx.meta_vars, inferred_typ, &t)?;
-        Ok(inferred_term)
-    }
-    /// The *inference* rule for calls is:
-    /// ```text
-    ///                 ...
-    ///           ──────────────────
-    ///            P, Γ ⊢ Cσ ⇒ ...
-    /// ```
-    fn infer(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
-        let Call { span, kind, name, args, .. } = self;
-
-        match kind {
-            CallKind::Codefinition | CallKind::Constructor => {
-                let Ctor { name, params, typ, .. } = &prg.ctor_or_codef(name, *span)?;
-                let args_out = check_args(args, prg, name, ctx, params, *span)?;
-                let typ_out = typ
-                    .subst_under_ctx(vec![params.len()].into(), &vec![args.args.clone()])
-                    .to_exp();
-                let typ_nf = typ_out.normalize(prg, &mut ctx.env())?;
-                Ok(Call {
-                    span: *span,
-                    kind: *kind,
-                    name: name.clone(),
-                    args: args_out,
-                    inferred_type: Some(typ_nf),
-                })
-            }
-            CallKind::LetBound => {
-                let Let { name, params, typ, .. } = prg.top_level_let(name, *span)?;
-                let args_out = check_args(args, prg, name, ctx, params, *span)?;
-                let typ_out =
-                    typ.subst_under_ctx(vec![params.len()].into(), &vec![args.args.clone()]);
-                let typ_nf = typ_out.normalize(prg, &mut ctx.env())?;
-                Ok(Call {
-                    span: *span,
-                    kind: *kind,
-                    name: name.clone(),
-                    args: args_out,
-                    inferred_type: Some(typ_nf),
-                })
-            }
-        }
-    }
-}
-
-// DotCall
-//
-//
-
-impl CheckInfer for DotCall {
-    /// The *checking* rule for dotcalls is:
-    /// ```text
-    ///                 ...
-    ///           ──────────────────
-    ///            P, Γ ⊢ e.Dσ ⇐ τ
-    /// ```
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        let inferred_term = self.infer(prg, ctx)?;
-        let inferred_typ = inferred_term.typ().ok_or(TypeError::Impossible {
-            message: "Expected inferred type".to_owned(),
-            span: None,
-        })?;
-        convert(ctx.levels(), &mut ctx.meta_vars, inferred_typ, &t)?;
-        Ok(inferred_term)
-    }
-
-    /// The *inference* rule for dotcalls is:
-    /// ```text
-    ///                 ...
-    ///           ──────────────────
-    ///            P, Γ ⊢ e.Dσ ⇒ ...
-    /// ```
-    fn infer(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
-        let DotCall { span, kind, exp, name, args, .. } = self;
-        let Dtor { name, params, self_param, ret_typ, .. } = &prg.dtor_or_def(name, *span)?;
-
-        let args_out = check_args(args, prg, name, ctx, params, *span)?;
-
-        let self_param_out = self_param
-            .typ
-            .subst_under_ctx(vec![params.len()].into(), &vec![args.args.clone()])
-            .to_exp();
-        let self_param_nf = self_param_out.normalize(prg, &mut ctx.env())?;
-
-        let exp_out = exp.check(prg, ctx, self_param_nf)?;
-
-        let subst = vec![args.args.clone(), vec![exp.clone()]];
-        let typ_out = ret_typ.subst_under_ctx(vec![params.len(), 1].into(), &subst);
-        let typ_out_nf = typ_out.normalize(prg, &mut ctx.env())?;
-
-        Ok(DotCall {
-            span: *span,
-            kind: *kind,
-            exp: exp_out,
-            name: name.clone(),
-            args: args_out,
-            inferred_type: Some(typ_out_nf),
-        })
-    }
-}
-
-// Anno
-//
-//
-
-impl CheckInfer for Anno {
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        let inferred_term = self.infer(prg, ctx)?;
-        let inferred_typ = inferred_term.typ().ok_or(TypeError::Impossible {
-            message: "Expected inferred type".to_owned(),
-            span: None,
-        })?;
-        convert(ctx.levels(), &mut ctx.meta_vars, inferred_typ, &t)?;
-        Ok(inferred_term)
-    }
-
-    /// The *inference* rule for type annotations is:
-    /// ```text
-    ///            P, Γ ⊢ τ ⇐ Type
-    ///            P, Γ ⊢ τ ▷ τ'
-    ///            P, Γ ⊢ e ⇐ τ'
-    ///           ──────────────────────
-    ///            P, Γ ⊢ (e : τ) ⇒ τ'
-    /// ```
-    fn infer(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
-        let Anno { span, exp, typ, .. } = self;
-        let typ_out = typ.check(prg, ctx, Rc::new(TypeUniv::new().into()))?;
-        let typ_nf = typ.normalize(prg, &mut ctx.env())?;
-        let exp_out = (**exp).check(prg, ctx, typ_nf.clone())?;
-        Ok(Anno { span: *span, exp: Rc::new(exp_out), typ: typ_out, normalized_type: Some(typ_nf) })
-    }
-}
-
-// TypeUniv
-//
-//
-
-impl CheckInfer for TypeUniv {
-    /// The *checking* rule for the type universe is:
-    /// ```text
-    ///            P, Γ ⊢ τ ≃ Type
-    ///           ──────────────────
-    ///            P, Γ ⊢ Type ⇐ τ
-    /// ```
-    fn check(&self, _prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        convert(ctx.levels(), &mut ctx.meta_vars, Rc::new(TypeUniv::new().into()), &t)?;
-        Ok(self.clone())
-    }
-
-    /// The *inference* rule for the type universe is:
-    /// ```text
-    ///           ─────────────────────
-    ///            P, Γ ⊢ Type ⇒ Type
-    /// ```
-    /// Note: The type universe is impredicative and the theory
-    /// therefore inconsistent.
-    fn infer(&self, _prg: &Module, _ctx: &mut Ctx) -> Result<Self, TypeError> {
-        Ok(self.clone())
-    }
-}
-
-// Hole
-//
-//
-
-impl CheckInfer for Hole {
-    fn check(&self, prg: &Module, ctx: &mut Ctx, t: Rc<Exp>) -> Result<Self, TypeError> {
-        let Hole { span, metavar, args, .. } = self;
-        let args: Vec<Vec<Rc<Exp>>> = args
-            .iter()
-            .map(|subst| subst.iter().map(|exp| exp.infer(prg, ctx)).collect::<Result<Vec<_>, _>>())
-            .collect::<Result<_, _>>()?;
-        Ok(Hole {
-            span: *span,
-            metavar: *metavar,
-            inferred_type: Some(t.clone()),
-            inferred_ctx: Some(ctx.vars.clone()),
-            args,
-        })
-    }
-
-    fn infer(&self, _prg: &Module, _ctx: &mut Ctx) -> Result<Self, TypeError> {
-        Err(TypeError::CannotInferHole { span: self.span().to_miette() })
-    }
-}
 
 // LocalMatch
 //
@@ -916,39 +540,6 @@ fn check_cocase(
         },
         *span,
     )
-}
-
-fn check_args(
-    this: &Args,
-    prg: &Module,
-    name: &str,
-    ctx: &mut Ctx,
-    params: &Telescope,
-    span: Option<Span>,
-) -> Result<Args, TypeError> {
-    if this.len() != params.len() {
-        return Err(TypeError::ArgLenMismatch {
-            name: name.to_owned(),
-            expected: params.len(),
-            actual: this.len(),
-            span: span.to_miette(),
-        });
-    }
-
-    let Telescope { params } =
-        params.subst_in_telescope(LevelCtx::empty(), &vec![this.args.clone()]);
-
-    let args = this
-        .args
-        .iter()
-        .zip(params)
-        .map(|(exp, Param { typ, .. })| {
-            let typ = typ.normalize(prg, &mut ctx.env())?;
-            exp.check(prg, ctx, typ)
-        })
-        .collect::<Result<_, _>>()?;
-
-    Ok(Args { args })
 }
 
 impl CheckTelescope for TelescopeInst {
